Add interface file for making main file signle responsibility
